requirements.txt
opencv-python~=4.9.0.80
numpy~=1.26.4
matplotlib~=3.9.0
Flask~=3.0.3
pillow~=10.3.0